Fix gcc3.x compiler warning (uint8_t value is always <= 255).

+++ src/inet_pton.c
@@ -92,7 +92,7 @@ inet_pton4(src, dst)
 		const char *pch;
 
 		if ((pch = (char *) strchr(digits, ch)) != NULL) {
-			uint8_t new = *tp * 10 + (pch - digits);
+			uint32_t new = *tp * 10 + (pch - digits);
 
 			if (new > 255)
 				return (0);
